add a go.mod
To make it possible to fetch the beats dependency, as that's
only published as a module now.

Eventually we'd want to migrate all dependency information from
glide.lock to go.mod

Unfortunately 'go get' doesn't take into account transitive
replacements, so we need to include the replacements required
by beats to make the build succeed on Travis.